Testing and logging
Related Work Items: #51, #55

+++ FresnoSolution/LanterneRouge.Fresno.Data/DataAccess/Entities/Measurement.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
 
-        public byte HeartRate { get; set; }
+        public int HeartRate { get; set; }
 
         public float Lactate { get; set; }
